Move TextSelectionMenu into /components/control
